+++ vite.config.ts
@@ -1,6 +1,16 @@
 import { sveltekit } from '@sveltejs/kit/vite'
-import { defineConfig } from 'vite'
+import { defineConfig } from 'vitest/config'
+import { svelteTesting } from '@testing-library/svelte/vite'
 
 export default defineConfig({
-  plugins: [sveltekit()]
+  define: {
+    __SVELTEKIT_PAYLOAD__: {
+      data: {}
+    }
+  },
+  plugins: [sveltekit(), svelteTesting()],
+  test: {
+    environment: 'jsdom',
+    setupFiles: ['./tests/unit/setup.ts'],
+  }
 })